LayoutTests/imported/w3c/web-platform-tests/css/css-contain/content-visibility/content-visibility-072-expected.txt


PASS one.getBoundingClientRect():
FAIL two.getBoundingClientRect():  assert_equals: y expected 3007 but got 3041
FAIL three.getBoundingClientRect():  assert_equals: y expected 3007 but got 3041
FAIL four.getBoundingClientRect():  assert_equals: y expected 3007 but got 3041
FAIL five.getBoundingClientRect():  assert_equals: y expected 3007 but got 3041
